Amit_ai_diploma/B2B/requirements.txt
# Test Automation Framework Requirements
# Core Testing Libraries
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1

# Reporting
allure-pytest==2.13.2
allure-python-commons==2.13.2

# Data Generation
faker==20.1.0

# Environment Management
python-dotenv==1.0.0

# Utilities
requests==2.31.0
selenium==4.15.2

# Development Tools
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Type Hints Support
typing-extensions==4.8.0
